Add unit test methods

+++ OOP/SerializedData/SerializedData/CreateFile.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace SerializedData
 {
-    class CreateFile:MessageSender
+    public class CreateFile:MessageSender
     {
         public string Path { get; set; }
         public string Text { get; set; }
@@ -13,19 +13,26 @@ namespace SerializedData
         {
             string file = Path + FileName + ".json";
             try{
-                if (File.Exists(file))
+                if (Text != "")
                 {
-                    File.Delete(file);
+                    if (File.Exists(file))
+                    {
+                        File.Delete(file);
+                    }
+                    using (FileStream fs = File.Create(file))
+                    {
+                        byte[] text = new UTF8Encoding(true).GetBytes(Text);
+                        fs.Write(text, 0, text.Length);
+                    }
+                    Console.WriteLine();
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine("Файл {0}.json успешно создан", FileName);
+                    Console.ResetColor();
                 }
-                using (FileStream fs = File.Create(file))
+                else
                 {
-                    byte[] text = new UTF8Encoding(true).GetBytes(Text);
-                    fs.Write(text, 0, text.Length);
+                    MessageSend("Red","Невозможно создать файл");
                 }
-                Console.WriteLine("Файл {0}.json успешно создан", FileName);
-                //TODO:
-                //Right method for MessageSend
-                //MessageSend("Green", "Файл {0}.json успешно создан", FileName);
             }
             catch(Exception e)
             {
